requirements.txt
Django==1.6.5
django-treemenus==0.9.1
django-ordered-model==0.4.0
